+++ packages/koot-webpack/loaders/react-hot/index.js
@@ -1,4 +1,6 @@
 const { getOptions } = require('loader-utils');
+const fs = require('fs-extra');
+const path = require('path');
 
 // modify based on:
 // https://github.com/NoamELB/react-hot-loader-loader
@@ -91,8 +93,16 @@ function transformSourceForExportClassDirectly(source) {
 
 // getters
 
+const customHotLoaderAbsolutePath = path.resolve(
+    __dirname,
+    '../../../koot/React/hoc/dev-hot.js'
+);
+const customHotLoader = fs.existsSync(customHotLoaderAbsolutePath)
+    ? customHotLoaderAbsolutePath.replace(/\\/g, '\\\\')
+    : 'koot/React/hoc/dev-hot';
 function getImportLine() {
-    return `import { hot } from 'react-hot-loader';\n`;
+    // return `import { hot } from 'react-hot-loader';\n`;
+    return `import hot from '${customHotLoader}';\n`;
 }
 
 function getExportDefaultClassName(source) {