+++ packages/koot/React/root.jsx
@@ -1,3 +1,4 @@
+// import { hot } from 'react-hot-loader/root';
 import React from 'react';
 import { Provider } from 'react-redux';
 import Router from 'react-router/lib/Router';
@@ -35,6 +36,7 @@ class Root extends React.Component {
     }
 }
 export default Root;
+// export default hot(Root);
 
 // let e = Root;