+++ test/projects/simple/src/views/home/index.jsx
@@ -5,6 +5,7 @@ import clientUpdatePageinfo from 'koot/utils/client-update-pageinfo';
 import Issue68 from '@components/issue-68';
 import TestAsyncFunction from './test-async-function';
 import TestGetPort from './test-get-port';
+import TestExtendForwardRef from './test-extend-forwardref';
 
 import styles from './styles.component.less';
 
@@ -43,6 +44,7 @@ const PageHome = ({ className }) => {
 
             <TestAsyncFunction />
             <TestGetPort />
+            <TestExtendForwardRef />
         </div>
     );
 };