+++ test/projects/simple/src/views/home/test-extend-forwardref/extended.jsx
@@ -1,0 +1,26 @@
+import React from 'react';
+import { extend } from 'koot';
+
+import styles from './extended.module.less';
+
+// Functional Component =======================================================
+
+const Extended = extend({
+    connect: true,
+    styles
+})(
+    React.memo(({ forwardedRef, className }) => {
+        // console.log('forwardedRef', forwardedRef);
+        return (
+            <div
+                ref={forwardedRef}
+                className={className}
+                id="__test-extend_forward_ref"
+            >
+                Extended
+            </div>
+        );
+    })
+);
+
+export default Extended;